Commit what I have for function parameter count matching. Causing segfault for tests. Need to figure out what is going on.

+++ src/Symbol.cpp
@@ -148,7 +148,7 @@ namespace Meow
 		return m_initialized;
 	}
 
-	int Symbol::getParameterCount() const
+	size_t Symbol::getParameterCount() const
 	{
 		return m_parameters.size();
 	}
